refactor(cli): move the product linter into scripts/lib so check-product.js stays a thin gate
`scripts/check-product.js` had grown to 706 lines — past the 700-line hard limit
in `.claude/rules/coding-standards.md`, which binds under `scripts/`. The fix is
the repo's own shape, not shorter prose: `check-doc-refs.js` is a thin runner
over `scripts/lib/doc-refs.js`, and the product linter now matches it.

Pure move, no rule changes:

- `scripts/lib/product-lint.js` (658 lines) takes the parsers, the C1–C7 checks,
  `validateProduct` and its helpers, plus the docblock that documents them. The
  body is byte-identical to the lines it came from; only the docblock's first
  line and its tier line are new (`Library tier:` — the module is pure, so `fs`,
  `path` and the `product/` path constants stayed behind).
- `scripts/check-product.js` (69 lines) keeps `readProduct()` and `main()`: it
  reads the corpus, calls the library, prints the report, exits 0/1. Its
  `module.exports` is gone — the suite now requires the library directly.
- `tests/scripts/check-product.test.js` keeps all 100 tests; only its require
  path changed. That suite is the falsifiability evidence for the gate, so it
  moves with the code rather than being trimmed with it.

Verified byte-identical on both paths: `node scripts/check-product.js` before
and after produces the same stdout, stderr and exit code on the live corpus
(exit 0) and on a mutated corpus that trips C7 (exit 1). The new file lands
inside jest's 80% line floor over `scripts/lib/` at 97.47%, and the global line
number rose 84.76% → 85.27%.

Gate: lint:fix, npm test (all 5 runners), check:docs, check:product,
check:cli-consumers, check:versions, check:hooks, check:eval-targets — all green.
